refactor: вынести в отдельные пакеты модели, сервисы и репозитории
