+++ src/main/java/ru/practicum/shareit/booking/model/Status.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.booking;
+package ru.practicum.shareit.booking.model;
 
 public enum Status {
     WAITING,
