+++ src/main/java/ru/practicum/shareit/exception/AlreadyExistException.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.utils.exception;
+package ru.practicum.shareit.exception;
 
 public class AlreadyExistException extends RuntimeException {
     public AlreadyExistException(String s) {
