+++ src/main/java/ru/practicum/shareit/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.utils.exception;
+package ru.practicum.shareit.exception;
 
 public class NotFoundException extends RuntimeException {
     public NotFoundException(String message) {
