+++ src/main/java/ru/practicum/shareit/exception/ShareitExceptionHandler.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.utils.exception;
+package ru.practicum.shareit.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
