+++ src/main/java/ru/practicum/shareit/exception/ValidationException.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.utils.exception;
+package ru.practicum.shareit.exception;
 
 public class ValidationException extends RuntimeException {
     public ValidationException(String message) {
