+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -3,9 +3,9 @@ package ru.practicum.shareit.item.dto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import ru.practicum.shareit.utils.validation.CreateValidation;
-import ru.practicum.shareit.utils.validation.NullOrNotBlank;
-import ru.practicum.shareit.utils.validation.UpdateValidation;
+import ru.practicum.shareit.validation.CreateValidation;
+import ru.practicum.shareit.validation.NullOrNotBlank;
+import ru.practicum.shareit.validation.UpdateValidation;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
