+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.item;
+package ru.practicum.shareit.item.service;
 
 import ru.practicum.shareit.item.model.Item;
