+++ src/main/java/ru/practicum/shareit/user/repository/UserRepository.java
@@ -1,8 +1,10 @@
-package ru.practicum.shareit.user;
+package ru.practicum.shareit.user.repository;
+
+import ru.practicum.shareit.user.model.User;
 
 import java.util.List;
 
-public interface UserService {
+public interface UserRepository {
     User get(Long id);
 
     List<User> getAll();
@@ -12,5 +14,4 @@ public interface UserService {
     User update(User user);
 
     boolean delete(Long id);
-
 }
