--- src/main/java/ru/practicum/shareit/utils/validation/UpdateValidation.java
@@ -1,4 +1,0 @@
-package ru.practicum.shareit.utils.validation;
-
-public interface UpdateValidation {
-}
