+++ src/main/java/ru/practicum/shareit/validation/CreateValidation.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.validation;
+
+public interface CreateValidation {
+}
