+++ src/main/java/ru/practicum/shareit/validation/UpdateValidation.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.validation;
+
+public interface UpdateValidation {
+}
