+++ src/test/java/ru/practicum/shareit/AbstractValidationTest.java
@@ -5,7 +5,7 @@ import javax.validation.Validator;
 import javax.validation.ValidatorFactory;
 
 public abstract class AbstractValidationTest {
-    public Validator validator;
+    protected Validator validator;
 
     public AbstractValidationTest() {
         ValidatorFactory validatorFactory = Validation.buildDefaultValidatorFactory();
