+++ src/test/java/ru/practicum/shareit/user/UserValidationTest.java
@@ -3,8 +3,8 @@ package ru.practicum.shareit.user;
 import org.junit.jupiter.api.Test;
 import ru.practicum.shareit.AbstractValidationTest;
 import ru.practicum.shareit.user.dto.UserDto;
-import ru.practicum.shareit.utils.validation.CreateValidation;
-import ru.practicum.shareit.utils.validation.UpdateValidation;
+import ru.practicum.shareit.validation.CreateValidation;
+import ru.practicum.shareit.validation.UpdateValidation;
 
 import javax.validation.ConstraintViolation;
 import java.util.Set;
